completion and signature help ordering (#33196)
