--- src/EditorFeatures/Core/CommandHandlers/IntelliSenseCommandHandler.cs
@@ -1,27 +1,0 @@
-﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
-
-using System;
-using System.ComponentModel.Composition;
-using Microsoft.CodeAnalysis.Host.Mef;
-using Microsoft.VisualStudio.Utilities;
-using VSCommanding = Microsoft.VisualStudio.Commanding;
-
-namespace Microsoft.CodeAnalysis.Editor.CommandHandlers
-{
-    [Export]
-    [Export(typeof(VSCommanding.ICommandHandler))]
-    [ContentType(ContentTypeNames.RoslynContentType)]
-    [Name(PredefinedCommandHandlerNames.IntelliSense)]
-    internal sealed class IntelliSenseCommandHandler : AbstractIntelliSenseCommandHandler
-    {
-        [ImportingConstructor]
-        [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-        public IntelliSenseCommandHandler(
-               CompletionCommandHandler completionCommandHandler,
-               SignatureHelpCommandHandler signatureHelpCommandHandler)
-            : base(completionCommandHandler,
-                   signatureHelpCommandHandler)
-        {
-        }
-    }
-}
