+++ src/EditorFeatures/Core/CommandHandlers/SignatureHelpBeforeCompletionCommandHandler.cs
@@ -3,17 +3,13 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.Composition;
-using System.Linq;
-using Microsoft.CodeAnalysis.Editor.Implementation.IntelliSense.SignatureHelp;
-using Microsoft.CodeAnalysis.Editor.Options;
-using Microsoft.CodeAnalysis.Editor.Shared.Extensions;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.Host.Mef;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
-using Microsoft.CodeAnalysis.Shared.Utilities;
 using Microsoft.CodeAnalysis.SignatureHelp;
 using Microsoft.VisualStudio.Commanding;
 using Microsoft.VisualStudio.Language.Intellisense;
+using Microsoft.VisualStudio.Language.Intellisense.AsyncCompletion;
 using Microsoft.VisualStudio.Text.Editor.Commanding;
 using Microsoft.VisualStudio.Text.Editor.Commanding.Commands;
 using Microsoft.VisualStudio.Utilities;
@@ -21,62 +17,44 @@ using VSCommanding = Microsoft.VisualStudio.Commanding;
 
 namespace Microsoft.CodeAnalysis.Editor.CommandHandlers
 {
+    /// <summary>
+    /// There are two forms of intellisense that may be active at the same time. Completion and
+    /// SigHelp. Completion precedes SigHelp in the command handler because it wants to make sure
+    /// it's operating on a buffer *after* Completion has changed it. i.e. if "WriteL(" is typed,
+    /// sig help wants to allow completion to complete that to "WriteLine(" before it tried to
+    /// proffer sig help. If we were to reverse things, then we'd get a bogus situation where sig
+    /// help would see "WriteL(" would have nothing to offer and would return.
+    /// 
+    /// However, despite wanting sighelp to receive typechar first and then defer it to completion,
+    /// we want completion to receive other events first (like escape, and navigation keys). We
+    /// consider completion to have higher priority for those commands. In order to accomplish that,
+    /// we introduced <see cref="SignatureHelpAfterCompletionCommandHandler"/>
+    /// This command handler then delegates escape, up and down to those command handlers. 
+    /// It is called after <see cref="PredefinedCommandHandlerNames.Completion"/> or <see cref="PredefinedCompletionNames.CompletionCommandHandler"/>
+    /// depending on the completion implemenetation.
+    /// </summary>
     [Export]
     [Export(typeof(VSCommanding.ICommandHandler))]
     [ContentType(ContentTypeNames.RoslynContentType)]
-    [Name(PredefinedCommandHandlerNames.SignatureHelp)]
+    [Name(PredefinedCommandHandlerNames.SignatureHelpBeforeCompletion)]
     [Order(Before = PredefinedCommandHandlerNames.Completion)]
-    internal class SignatureHelpCommandHandler :
-        ForegroundThreadAffinitizedObject,
+    [Order(Before = PredefinedCompletionNames.CompletionCommandHandler)]
+    internal class SignatureHelpBeforeCompletionCommandHandler :
+        AbstractSignatureHelpCommandHandler,
         IChainedCommandHandler<TypeCharCommandArgs>,
         IChainedCommandHandler<InvokeSignatureHelpCommandArgs>
     {
-        private readonly IIntelliSensePresenter<ISignatureHelpPresenterSession, ISignatureHelpSession> _signatureHelpPresenter;
-        private readonly IAsynchronousOperationListener _listener;
-        private readonly IList<Lazy<ISignatureHelpProvider, OrderableLanguageMetadata>> _signatureHelpProviders;
-
         public string DisplayName => EditorFeaturesResources.Signature_Help;
 
         [ImportingConstructor]
         [Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-        public SignatureHelpCommandHandler(
+        public SignatureHelpBeforeCompletionCommandHandler(
             IThreadingContext threadingContext,
             [ImportMany] IEnumerable<Lazy<ISignatureHelpProvider, OrderableLanguageMetadata>> signatureHelpProviders,
             [ImportMany] IEnumerable<Lazy<IIntelliSensePresenter<ISignatureHelpPresenterSession, ISignatureHelpSession>, OrderableMetadata>> signatureHelpPresenters,
             IAsynchronousOperationListenerProvider listenerProvider)
-            : base(threadingContext)
+            : base(threadingContext, signatureHelpProviders, signatureHelpPresenters, listenerProvider)
         {
-            _signatureHelpProviders = ExtensionOrderer.Order(signatureHelpProviders);
-            _listener = listenerProvider.GetListener(FeatureAttribute.SignatureHelp);
-            _signatureHelpPresenter = ExtensionOrderer.Order(signatureHelpPresenters).Select(lazy => lazy.Value).FirstOrDefault();
-        }
-
-        private bool TryGetController(EditorCommandArgs args, out Controller controller)
-        {
-            AssertIsForeground();
-
-            // If args is `InvokeSignatureHelpCommandArgs` then sig help was explicitly invoked by the user and should
-            // be shown whether or not the option is set.
-            if (!(args is InvokeSignatureHelpCommandArgs) && !args.SubjectBuffer.GetFeatureOnOffOption(SignatureHelpOptions.ShowSignatureHelp))
-            {
-                controller = null;
-                return false;
-            }
-
-            // If we don't have a presenter, then there's no point in us even being involved.  Just
-            // defer to the next handler in the chain.
-            if (_signatureHelpPresenter == null)
-            {
-                controller = null;
-                return false;
-            }
-
-            controller = Controller.GetInstance(
-                ThreadingContext,
-                args, _signatureHelpPresenter,
-                _listener, _signatureHelpProviders);
-
-            return true;
         }
 
         private bool TryGetControllerCommandHandler<TCommandArgs>(TCommandArgs args, out VSCommanding.ICommandHandler commandHandler)
@@ -143,36 +121,6 @@ namespace Microsoft.CodeAnalysis.Editor.CommandHandlers
         {
             AssertIsForeground();
             ExecuteCommandWorker(args, nextHandler, context);
-        }
-
-        internal bool TryHandleEscapeKey(EscapeKeyCommandArgs commandArgs)
-        {
-            if (!TryGetController(commandArgs, out var controller))
-            {
-                return false;
-            }
-
-            return controller.TryHandleEscapeKey();
-        }
-
-        internal bool TryHandleUpKey(UpKeyCommandArgs commandArgs)
-        {
-            if (!TryGetController(commandArgs, out var controller))
-            {
-                return false;
-            }
-
-            return controller.TryHandleUpKey();
-        }
-
-        internal bool TryHandleDownKey(DownKeyCommandArgs commandArgs)
-        {
-            if (!TryGetController(commandArgs, out var controller))
-            {
-                return false;
-            }
-
-            return controller.TryHandleDownKey();
         }
     }
 }
