+++ src/EditorFeatures/Core/Extensibility/Commands/PredefinedCommandHandlerNames.cs
@@ -42,8 +42,7 @@ namespace Microsoft.CodeAnalysis.Editor
         public const string Commit = "Commit Command Handler";
 
         /// <summary>
-        /// Command handler name for Completion. Some additional Completion commands are handled by
-        /// the <see cref="IntelliSense"/> command handler.
+        /// Command handler name for Completion.
         /// </summary>
         public const string Completion = "Completion Command Handler";
 
@@ -108,12 +107,6 @@ namespace Microsoft.CodeAnalysis.Editor
         public const string Indent = "Indent Command Handler";
 
         /// <summary>
-        /// Command handler name for IntelliSense. This command handler handles some commands for
-        /// <see cref="Completion"/>, <see cref="QuickInfo"/>, and <see cref="SignatureHelp"/>.
-        /// </summary>
-        public const string IntelliSense = nameof(IntelliSense);
-
-        /// <summary>
         /// Command handler name for Navigate to Highlighted Reference.
         /// </summary>
         public const string NavigateToHighlightedReference = "Navigate to Highlighted Reference Command Handler";
@@ -124,8 +117,7 @@ namespace Microsoft.CodeAnalysis.Editor
         public const string OrganizeDocument = "Organize Document Command Handler";
 
         /// <summary>
-        /// Command handler name for Quick Info. Some additional Quick Info commands are handled by
-        /// the <see cref="IntelliSense"/> command handler.
+        /// Command handler name for Quick Info.
         /// </summary>
         public const string QuickInfo = "Quick Info Command Handler";
 
@@ -140,10 +132,14 @@ namespace Microsoft.CodeAnalysis.Editor
         public const string RenameTrackingCancellation = "Rename Tracking Cancellation Command Handler";
 
         /// <summary>
-        /// Command handler name for Signature Help. Some additional Signature Help commands are
-        /// handled by the <see cref="IntelliSense"/> command handler.
+        /// Command handler name for a Signature Help command handler executing before <see cref="Completion"/>.
         /// </summary>
-        public const string SignatureHelp = "Signature Help Command Handler";
+        public const string SignatureHelpBeforeCompletion = "Signature Help Before Completion Command Handler";
+
+        /// <summary>
+        /// Command handler name for a Signature Help command handler executing after <see cref="Completion"/>.
+        /// </summary>
+        public const string SignatureHelpAfterCompletion = "Signature Help After Completion Command Handler";
 
         /// <summary>
         /// Command handler name for Paste Content in Interactive Format.
